pyproject.toml
[project]
name = "chess-engine"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-chess >=1.999,<2.0.0",
    "torch >=2.6.0,<3.0.0",
    "pytest >=8.3.4,<9.0.0",
    "lightning >=2.4.0,<3.0.0",
    "python-dotenv >=1.0.1,<2.0.0",
    "numpy >=2.0.1,<3.0.0",
    "isort>=6.0.1",
    "pre-commit>=4.1.0",
]

[dependency-groups]
training = [
    "wandb >=0.19.1,<0.20.0",
    "marimo >=0.11.8,<0.12.0",
    "altair >=5.5.0,<6.0.0",
    "pandas >=2.2.3,<3.0.0",
    "nbformat >=5.10.4,<6.0.0",
    "watchdog>=6.0.0",
    "tqdm>=4.65.0,<5.0.0",
]

[tool.uv]
default-groups = ["training"]

[tool.marimo.runtime]
output_max_bytes = 20_000_000

[tool.isort]
profile = "black"
src_paths = ["src", "tests"]

[project.scripts]
engine = "src.engine.engine:main"
train_debug = "src.training.train:main_debug"
train_pico = "src.training.train:main_pico"
lc0_to_npz = "src.preprocessing.lc0_to_npz:main"
lc0_explorer = "src.preprocessing.lc0_data_explorer:main"
npz_explorer = "src.preprocessing.npz_explorer:main"


[tool.hatch.build.targets.wheel]
packages = ["src"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
